Big refactor: Move a bunch of functions out of LegendList into LegendListHelpers for a small perf benefit and to ensure we don't accidentally access state data
