+++ src/LegendList.tsx
@@ -1,20 +1,25 @@
 import * as React from 'react';
-import { ForwardedRef, forwardRef, ReactElement, useCallback, useEffect, useMemo, useRef } from 'react';
+import { ForwardedRef, forwardRef, ReactElement, useMemo, useRef } from 'react';
+import { Dimensions, ScrollView, StyleSheet } from 'react-native';
+import { DEFAULT_SCROLL_BUFFER } from './constants';
 import {
-    Dimensions,
-    LayoutChangeEvent,
-    NativeScrollEvent,
-    NativeSyntheticEvent,
-    ScrollView,
-    StyleSheet,
-    unstable_batchedUpdates,
-} from 'react-native';
+    allocateContainers,
+    applyDefaultProps,
+    calculateInitialOffset as calculateInitialOffsetHelper,
+    calculateItemsInView as calculateItemsInViewHelper,
+    checkAtBottom as checkAtBottomHelper,
+    getId as getIdHelper,
+    getItemSize as getItemSizeHelper,
+    getRenderedItem as getRenderedItemHelper,
+    handleScrollDebounced as handleScrollDebouncedHelper,
+    handleScroll as handleScrollHelper,
+    onLayout as onLayoutHelper,
+    setTotalLength,
+    updateItemLength as updateItemLengthHelper,
+} from './LegendListHelpers';
 import { ListComponent } from './ListComponent';
-import { peek$, set$, StateProvider, useStateContext } from './state';
-import type { LegendListProps } from './types';
-
-const DEFAULT_SCROLL_BUFFER = 0;
-const POSITION_OUT_OF_VIEW = -10000;
+import { set$, StateProvider, useStateContext } from './state';
+import type { InternalState, LegendListProps } from './types';
 
 export const LegendList: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<ScrollView> }) => ReactElement =
     forwardRef(function LegendList<T>(props: LegendListProps<T>, forwardedRef: ForwardedRef<ScrollView>) {
@@ -26,7 +31,8 @@ export const LegendList: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Sc
     }) as any;
 
 const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<ScrollView> }) => ReactElement = forwardRef(
-    function LegendListInner<T>(props: LegendListProps<T>, forwardedRef: ForwardedRef<ScrollView>) {
+    function LegendListInner<T>(props_: LegendListProps<T>, forwardedRef: ForwardedRef<ScrollView>) {
+        const props = applyDefaultProps(props_);
         const {
             data,
             initialScrollIndex,
@@ -34,89 +40,23 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
             horizontal,
             style: styleProp,
             contentContainerStyle: contentContainerStyleProp,
-            initialNumContainers,
             drawDistance,
-            recycleItems = true,
-            onEndReachedThreshold = 0.5,
-            maintainScrollAtEnd = false,
-            maintainScrollAtEndThreshold = 0.1,
-            alignItemsAtEnd = false,
-            onScroll: onScrollProp,
-            keyExtractor,
-            renderItem,
-            estimatedItemSize,
-            getEstimatedItemSize,
-            onEndReached,
-            onViewableRangeChanged,
             ...rest
         } = props;
 
         const ctx = useStateContext();
 
-        const internalRef = useRef<ScrollView>(null);
-        const refScroller = (forwardedRef || internalRef) as React.MutableRefObject<ScrollView>;
+        const refScroller = (forwardedRef || useRef<ScrollView>(null)) as React.MutableRefObject<ScrollView>;
         const scrollBuffer = drawDistance ?? DEFAULT_SCROLL_BUFFER;
         // Experimental: It works ok on iOS when scrolling up, but is doing weird things when sizes are changing.
         // And it doesn't work at all on Android because it uses contentInset. I'll try it again later.
         // Ideally it would work by adjusting the contentOffset but in previous attempts that was causing jitter.
         const supportsEstimationAdjustment = false; //   Platform.OS === "ios";
 
-        const styleFlattened = StyleSheet.flatten(styleProp);
-        const style = useMemo(() => styleFlattened, [JSON.stringify(styleProp)]);
-        const contentContainerStyleFlattened = StyleSheet.flatten(contentContainerStyleProp);
-        const contentContainerStyle = useMemo(
-            () => contentContainerStyleFlattened,
-            [JSON.stringify(contentContainerStyleProp)],
-        );
-
-        const refState = useRef<{
-            positions: Map<string, number>;
-            lengths: Map<String, number>;
-            pendingAdjust: number;
-            animFrame: number | null;
-            isStartReached: boolean;
-            isEndReached: boolean;
-            isAtBottom: boolean;
-            data: T[];
-            idsInFirstRender: Set<string>;
-            hasScrolled: boolean;
-            scrollLength: number;
-            startBuffered: number;
-            startNoBuffer: number;
-            endBuffered: number;
-            endNoBuffer: number;
-            scroll: number;
-            topPad: number;
-        }>();
-        const getId = (index: number): string => {
-            const data = refState.current?.data;
-            if (!data) {
-                return '';
-            }
-            const ret = index < data.length ? (keyExtractor ? keyExtractor(data[index], index) : index) : null;
-            return ret + '';
-        };
-
-        const getItemLength = (index: number, data: T) => {
-            return getEstimatedItemSize ? getEstimatedItemSize(index, data) : estimatedItemSize;
-        };
-        const calculateInitialOffset = () => {
-            if (initialScrollIndex) {
-                if (getEstimatedItemSize) {
-                    let offset = 0;
-                    for (let i = 0; i < initialScrollIndex; i++) {
-                        offset += getEstimatedItemSize(i, data[i]);
-                    }
-                    return offset;
-                } else if (estimatedItemSize) {
-                    return initialScrollIndex * estimatedItemSize;
-                }
-            }
-            return undefined;
-        };
+        const refState = useRef<InternalState<T>>();
 
         const initialContentOffset =
-            initialScrollOffset ?? useMemo(calculateInitialOffset, [initialScrollIndex, estimatedItemSize]);
+            initialScrollOffset ?? useMemo(calculateInitialOffsetHelper.bind(undefined, props), []);
 
         if (!refState.current) {
             refState.current = {
@@ -127,7 +67,6 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
                 isStartReached: false,
                 isEndReached: false,
                 isAtBottom: false,
-                data: data,
                 idsInFirstRender: undefined as any,
                 hasScrolled: false,
                 scrollLength: Dimensions.get('window')[horizontal ? 'width' : 'height'],
@@ -137,219 +76,59 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
                 endNoBuffer: 0,
                 scroll: initialContentOffset || 0,
                 topPad: 0,
+                previousViewableItems: new Set(),
+                props,
+                ctx,
+                scrollBuffer,
             };
-            refState.current.idsInFirstRender = new Set(data.map((_: any, i: number) => getId(i)));
+            refState.current.idsInFirstRender = new Set(
+                data.map((_: any, i: number) => getIdHelper(refState.current!, i)),
+            );
         }
-        refState.current.data = data;
-        set$(ctx, `numItems`, data.length);
+        refState.current.props = props;
+        refState.current.ctx = ctx;
 
-        const setTotalLength = (length: number) => {
-            set$(ctx, `totalLength`, length);
-            const screenLength = refState.current!.scrollLength;
-            if (alignItemsAtEnd) {
-                const listPaddingTop =
-                    ((style as any)?.paddingTop || 0) + ((contentContainerStyle as any)?.paddingTop || 0);
-                set$(ctx, `paddingTop`, Math.max(0, screenLength - length - listPaddingTop));
-            }
-        };
-
-        const allocateContainers = useCallback(() => {
-            const scrollLength = refState.current!.scrollLength;
-            const averageItemSize = estimatedItemSize ?? getEstimatedItemSize?.(0, data[0]);
-            const numContainers =
-                initialNumContainers || Math.ceil((scrollLength + scrollBuffer * 2) / averageItemSize) + 4;
-
-            for (let i = 0; i < numContainers; i++) {
-                set$(ctx, `containerIndex${i}`, -1);
-                set$(ctx, `containerPosition${i}`, POSITION_OUT_OF_VIEW);
-            }
-
-            set$(ctx, `numContainers`, numContainers);
-        }, []);
-
-        const getRenderedItem = useCallback(
-            (index: number) => {
-                const data = refState.current?.data;
-                if (!data) {
-                    return null;
-                }
-                const renderedItem = renderItem?.({
-                    item: data[index],
-                    index,
-                } as any);
-
-                return renderedItem;
-            },
-            [renderItem],
+        const styleFlattened = StyleSheet.flatten(styleProp);
+        const style = useMemo(() => styleFlattened, [JSON.stringify(styleProp)]);
+        const contentContainerStyleFlattened = StyleSheet.flatten(contentContainerStyleProp);
+        const contentContainerStyle = useMemo(
+            () => contentContainerStyleFlattened,
+            [JSON.stringify(contentContainerStyleProp)],
         );
 
-        const calculateItemsInView = useCallback(() => {
-            // This should be a good optimization to make sure that all React updates happen in one frame
-            // but it should be tested more with and without it to see if it's better.
-            unstable_batchedUpdates(() => {
-                const {
-                    data,
-                    scrollLength,
-                    scroll: scrollState,
-                    topPad,
-                    startNoBuffer: startNoBufferState,
-                    startBuffered: startBufferedState,
-                    endNoBuffer: endNoBufferState,
-                    endBuffered: endBufferedState,
-                } = refState.current!;
-                if (!data) {
-                    return;
-                }
-                const scroll = scrollState - topPad;
+        // Create functions that are bound to the state to avoid re-creating them on every render.
+        // This should be a minor optimization when data is changing often. And putting them elsewhere
+        // makes sure we always get the latest values from state and avoid accidentally using stale values.
+        const fns = useMemo(
+            () => ({
+                getRenderedItem: getRenderedItemHelper.bind(undefined, refState.current!),
+                getId: getIdHelper.bind(undefined, refState.current!),
+                getItemSize: getItemSizeHelper.bind(undefined, refState.current!),
+                calculateItemsInView: calculateItemsInViewHelper.bind(undefined, refState.current!),
+                updateItemLength: updateItemLengthHelper.bind(undefined, refState.current!, refScroller),
+                checkAtBottom: checkAtBottomHelper.bind(undefined, refState.current!),
+                handleScrollDebounced: handleScrollDebouncedHelper.bind(undefined, refState.current!),
+                onLayout: onLayoutHelper.bind(undefined, refState.current!),
+                handleScroll: handleScrollHelper.bind(
+                    undefined,
+                    refState.current!,
+                    handleScrollDebouncedHelper.bind(undefined, refState.current!),
+                ),
+            }),
+            [],
+        );
+        const {
+            calculateItemsInView,
+            getId,
+            getItemSize,
+            checkAtBottom,
+            updateItemLength,
+            getRenderedItem,
+            onLayout,
+            handleScroll,
+        } = fns;
 
-                const { lengths, positions } = refState.current!;
-
-                let startNoBuffer: number | null = null;
-                let startBuffered: number | null = null;
-                let endNoBuffer: number | null = null;
-                let endBuffered: number | null = null;
-
-                // Go backwards from the last start position to find the first item that is in view
-                // This is an optimization to avoid looping through all items, which could slow down
-                // when scrolling at the end of a long list.
-                let loopStart = startBufferedState || 0;
-                if (startBufferedState) {
-                    for (let i = startBufferedState; i >= 0; i--) {
-                        const id = getId(i)!;
-                        const top = positions.get(id)!;
-                        if (top !== undefined) {
-                            const length = lengths.get(id) ?? getItemLength(i, data[i]);
-                            const bottom = top + length;
-                            if (bottom > scroll - scrollBuffer) {
-                                loopStart = i;
-                            } else {
-                                break;
-                            }
-                        }
-                    }
-                }
-
-                let top = loopStart > 0 ? positions.get(getId(loopStart))! : 0;
-
-                for (let i = loopStart; i < data!.length; i++) {
-                    const id = getId(i)!;
-                    const length = lengths.get(id) ?? getItemLength(i, data[i]);
-
-                    if (positions.get(id) !== top) {
-                        positions.set(id, top);
-                    }
-
-                    if (startNoBuffer === null && top + length > scroll) {
-                        startNoBuffer = i;
-                    }
-                    if (startBuffered === null && top + length > scroll - scrollBuffer) {
-                        startBuffered = i;
-                    }
-                    if (startNoBuffer !== null) {
-                        if (top <= scroll + scrollLength) {
-                            endNoBuffer = i;
-                        }
-                        if (top <= scroll + scrollLength + scrollBuffer) {
-                            endBuffered = i;
-                        } else {
-                            break;
-                        }
-                    }
-
-                    top += length;
-                }
-
-                Object.assign(refState.current!, {
-                    startBuffered,
-                    startNoBuffer,
-                    endBuffered,
-                    endNoBuffer,
-                });
-
-                if (startBuffered !== null && endBuffered !== null) {
-                    const prevNumContainers = ctx.values.get('numContainers');
-                    let numContainers = prevNumContainers;
-                    for (let i = startBuffered; i <= endBuffered; i++) {
-                        let isContained = false;
-                        // See if this item is already in a container
-                        for (let j = 0; j < numContainers; j++) {
-                            const index = peek$(ctx, `containerIndex${j}`);
-                            if (index === i) {
-                                isContained = true;
-                                break;
-                            }
-                        }
-                        // If it's not in a container, then we need to recycle a container out of view
-                        if (!isContained) {
-                            let didRecycle = false;
-                            for (let u = 0; u < numContainers; u++) {
-                                const index = peek$(ctx, `containerIndex${u}`);
-
-                                if (index < startBuffered || index > endBuffered) {
-                                    set$(ctx, `containerIndex${u}`, i);
-                                    didRecycle = true;
-                                    break;
-                                }
-                            }
-                            if (!didRecycle) {
-                                if (__DEV__) {
-                                    console.warn(
-                                        '[legend-list] No container to recycle, consider increasing initialContainers or estimatedItemLength',
-                                        i,
-                                    );
-                                }
-                                const id = numContainers;
-                                numContainers++;
-                                set$(ctx, `containerIndex${id}`, i);
-                                set$(ctx, `containerPosition${id}`, POSITION_OUT_OF_VIEW);
-                            }
-                        }
-                    }
-
-                    if (numContainers !== prevNumContainers) {
-                        set$(ctx, `numContainers`, numContainers);
-                    }
-
-                    // Update top positions of all containers
-                    // TODO: This could be optimized to only update the containers that have changed
-                    // but it likely would have little impact. Remove this comment if not worth doing.
-                    for (let i = 0; i < numContainers; i++) {
-                        const itemIndex = peek$(ctx, `containerIndex${i}`);
-                        const item = data[itemIndex];
-                        if (item) {
-                            const id = getId(itemIndex);
-                            if (itemIndex < startBuffered || itemIndex > endBuffered) {
-                                set$(ctx, `containerPosition${i}`, POSITION_OUT_OF_VIEW);
-                            } else {
-                                const pos = positions.get(id) ?? -1;
-                                const prevPos = peek$(ctx, `containerPosition${i}`);
-                                if (pos >= 0 && pos !== prevPos) {
-                                    set$(ctx, `containerPosition${i}`, pos);
-                                }
-                            }
-                        }
-                    }
-
-                    // TODO: Add the more complex onViewableItemsChanged
-                    if (onViewableRangeChanged) {
-                        if (
-                            startNoBuffer !== startNoBufferState ||
-                            startBuffered !== startBufferedState ||
-                            endNoBuffer !== endNoBufferState ||
-                            endBuffered !== endBufferedState
-                        ) {
-                            onViewableRangeChanged({
-                                start: startNoBuffer!,
-                                startBuffered,
-                                end: endNoBuffer!,
-                                endBuffered,
-                                items: data.slice(startNoBuffer!, endNoBuffer! + 1),
-                            });
-                        }
-                    }
-                }
-            });
-        }, [data]);
+        set$(ctx, `numItems`, data.length);
 
         // const adjustTopPad = (diff: number) => {
         //     // TODO: Experimental, find a better way to do this.
@@ -369,7 +148,7 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
         // };
 
         useMemo(() => {
-            allocateContainers();
+            allocateContainers(refState.current!);
             calculateItemsInView();
 
             // Set an initial total height based on what we know
@@ -377,28 +156,10 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
             let totalLength = 0;
             for (let i = 0; i < data.length; i++) {
                 const id = getId(i);
-                totalLength += lengths.get(id) ?? getItemLength(i, data[i]);
+                totalLength += lengths.get(id) ?? getItemSize(i, data[i]);
             }
-            setTotalLength(totalLength);
+            setTotalLength(refState.current!, totalLength);
         }, []);
-
-        const checkAtBottom = () => {
-            const { scrollLength, scroll } = refState.current!;
-            const totalLength = peek$(ctx, 'totalLength');
-            // Check if at end
-            const distanceFromEnd = totalLength - scroll - scrollLength;
-            if (refState.current) {
-                refState.current.isAtBottom = distanceFromEnd < scrollLength * maintainScrollAtEndThreshold;
-            }
-            if (onEndReached && !refState.current?.isEndReached) {
-                if (distanceFromEnd < onEndReachedThreshold! * scrollLength) {
-                    if (refState.current) {
-                        refState.current.isEndReached = true;
-                    }
-                    onEndReached({ distanceFromEnd });
-                }
-            }
-        };
 
         useMemo(() => {
             if (refState.current) {
@@ -407,113 +168,6 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
             calculateItemsInView();
             checkAtBottom();
         }, [data]);
-
-        const updateItemLength = useCallback((index: number, length: number) => {
-            const data = refState.current?.data;
-            if (!data) {
-                return;
-            }
-            const lengths = refState.current?.lengths!;
-            const id = getId(index);
-            const wasInFirstRender = refState.current?.idsInFirstRender.has(id);
-
-            const prevLength = lengths.get(id) || (wasInFirstRender ? getItemLength(index, data[index]) : 0);
-            // let scrollNeedsAdjust = 0;
-
-            if (!prevLength || prevLength !== length) {
-                // TODO: Experimental scroll adjusting
-                // const diff = length - (prevLength || 0);
-                // const startNoBuffer = visibleRange$.startNoBuffer.peek();
-                // if (refPositions.current?.hasScrolled && wasInFirstRender && index <= startNoBuffer) {
-                //     scrollNeedsAdjust += diff;
-                // }
-
-                lengths.set(id, length);
-                const totalLength = peek$(ctx, 'totalLength');
-                setTotalLength(totalLength + (length - prevLength));
-
-                if (refState.current?.isAtBottom && maintainScrollAtEnd) {
-                    // TODO: This kinda works, but with a flash. Since setNativeProps is less ideal we'll favor the animated one for now.
-                    // scrollRef.current?.setNativeProps({
-                    //   contentContainerStyle: {
-                    //     height:
-                    //       visibleRange$.totalLength.get() + visibleRange$.topPad.get() + 48,
-                    //   },
-                    //   contentOffset: {
-                    //     y:
-                    //       visibleRange$.totalLength.peek() +
-                    //       visibleRange$.topPad.peek() -
-                    //       SCREEN_LENGTH +
-                    //       48 * 3,
-                    //   },
-                    // });
-
-                    // TODO: This kinda works too, but with more of a flash
-                    requestAnimationFrame(() => {
-                        refScroller.current?.scrollToEnd({
-                            animated: true,
-                        });
-                    });
-                }
-
-                // TODO: Could this be optimized to only calculate items in view that have changed?
-
-                // Calculate positions if not currently scrolling and have a calculate already pending
-                if (!refState.current?.animFrame) {
-                    calculateItemsInView();
-                }
-
-                // TODO: Experimental
-                // if (scrollNeedsAdjust) {
-                //     adjustTopPad(scrollNeedsAdjust);
-                // }
-            }
-        }, []);
-
-        const handleScrollDebounced = useCallback(() => {
-            calculateItemsInView();
-            checkAtBottom();
-
-            // Reset the debounce
-            if (refState.current) {
-                refState.current.animFrame = null;
-            }
-        }, []);
-
-        const onLayout = useCallback((event: LayoutChangeEvent) => {
-            const scrollLength = event.nativeEvent.layout[horizontal ? 'width' : 'height'];
-            refState.current!.scrollLength = scrollLength;
-        }, []);
-
-        const handleScroll = useCallback(
-            (
-                event: {
-                    nativeEvent: NativeScrollEvent;
-                },
-                fromSelf?: boolean,
-            ) => {
-                // in some cases when we set ScrollView contentOffset prop, there comes event from with 0 height and width
-                // this causes blank list display, looks to be Paper implementation problem
-                // let's filter out such events
-                if (event.nativeEvent?.contentSize?.height === 0 && event.nativeEvent.contentSize?.width === 0) {
-                    return;
-                }
-                refState.current!.hasScrolled = true;
-                const newScroll = event.nativeEvent.contentOffset[horizontal ? 'x' : 'y'];
-                // Update the scroll position to use in checks
-                refState.current!.scroll = newScroll;
-
-                // Debounce a calculate if no calculate is already pending
-                if (refState.current && !refState.current.animFrame) {
-                    refState.current.animFrame = requestAnimationFrame(handleScrollDebounced);
-                }
-
-                if (!fromSelf) {
-                    onScrollProp?.(event as NativeSyntheticEvent<NativeScrollEvent>);
-                }
-            },
-            [],
-        );
 
         return (
             <ListComponent
@@ -527,9 +181,6 @@ const LegendListInner: <T>(props: LegendListProps<T> & { ref?: ForwardedRef<Scro
                 updateItemLength={updateItemLength}
                 handleScroll={handleScroll}
                 onLayout={onLayout}
-                recycleItems={recycleItems}
-                alignItemsAtEnd={alignItemsAtEnd}
-                estimatedItemSize={estimatedItemSize}
             />
         );
     },
