+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_SCROLL_BUFFER = 0;
+export const POSITION_OUT_OF_VIEW = -10000;
