+++ src/index.ts
@@ -1,2 +1,2 @@
 export { LegendList } from './LegendList';
-export type * from './types';
+export type { LegendListProps, LegendListRenderItemInfo, ViewableRange } from './types';
