+++ src/state.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import type { StateType, StateContext } from './types';
 
 // This is an implementation of a simple state management system, inspired by Legend State.
 // It stores values and listeners in Maps, with peek$ and set$ functions to get and set values.
@@ -8,20 +9,7 @@ import * as React from 'react';
 // once for each unique name. So we don't need to manage a Set of listeners or dispose them,
 // which saves needing useEffect hooks or managing listeners in a Set.
 
-export type ListenerType =
-    | 'numContainers'
-    | `containerIndex${number}`
-    | `containerPosition${number}`
-    | `numItems`
-    | 'totalLength'
-    | 'paddingTop';
-
-interface ListenerContext {
-    listeners: Map<ListenerType, () => void>;
-    values: Map<ListenerType, any>;
-}
-
-const ContextListener = React.createContext<ListenerContext | null>(null);
+const ContextListener = React.createContext<StateContext | null>(null);
 
 export function StateProvider({ children }: { children: React.ReactNode }) {
     const [value] = React.useState(() => ({
@@ -35,7 +23,7 @@ export function useStateContext() {
     return React.useContext(ContextListener)!;
 }
 
-export function use$<T>(signalName: ListenerType): T {
+export function use$<T>(signalName: StateType): T {
     const { listeners, values } = React.useContext(ContextListener)!;
     const [, forceUpdate] = React.useReducer((x) => x + 1, 0);
     listeners.set(signalName, forceUpdate);
@@ -43,12 +31,12 @@ export function use$<T>(signalName: ListenerType): T {
     return values.get(signalName);
 }
 
-export function peek$(ctx: ListenerContext, signalName: ListenerType) {
+export function peek$(ctx: StateContext, signalName: StateType) {
     const { values } = ctx;
     return values.get(signalName);
 }
 
-export function set$(ctx: ListenerContext, signalName: ListenerType, value: any) {
+export function set$(ctx: StateContext, signalName: StateType, value: any) {
     const { listeners, values } = ctx;
     if (values.get(signalName) !== value) {
         values.set(signalName, value);
